+++ src/main/groovy/org/grails/gradle/plugin/core/GrailsGradlePlugin.groovy
@@ -142,6 +142,8 @@ class GrailsGradlePlugin extends GroovyPlugin {
 
         configureRunCommand(project)
 
+        configureJavaTime(project)
+
         configurePathingJar(project)
     }
 
@@ -650,6 +652,31 @@ class GrailsGradlePlugin extends GroovyPlugin {
 
     protected FileCollection resolveClassesDirs(SourceSetOutput output, Project project) {
         output?.classesDirs ?: project.files(new File(project.buildDir, "classes/main"))
+    }
+
+    protected void configureJavaTime(Project project) {
+        GrailsExtension grailsExt = project.extensions.getByType(GrailsExtension)
+
+        project.tasks.withType(GroovyCompile).configureEach { groovyCompileTask ->
+            if (grailsExt.javaTime) {
+                groovyCompileTask.doFirst {
+                    def configScriptStream = getClass().getResourceAsStream("/GrailsCompilerConfig.groovy")
+                    if (configScriptStream != null) {
+                        def tempConfigScriptFile = File.createTempFile("build/GrailsCompilerConfig", ".groovy")
+                        tempConfigScriptFile.mkdirs()
+                        tempConfigScriptFile.deleteOnExit()
+
+                        def existingScript = groovyCompileTask.groovyOptions.configurationScript
+                        if (existingScript) {
+                            tempConfigScriptFile << existingScript.text
+                        }
+
+                        tempConfigScriptFile.text = configScriptStream.text
+                        groovyCompileTask.groovyOptions.configurationScript = tempConfigScriptFile
+                    }
+                }
+            }
+        }
     }
 
     @CompileDynamic
